make inference library
