+++ inference/ds_zero/model.py
@@ -1,36 +1,66 @@
+import os
 from argparse import Namespace
 from typing import List, Union
 
+import deepspeed
 import torch
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
+from transformers.deepspeed import HfDeepSpeedConfig
 
-import utils
-from utils import (
-    Model,
-    benchmark_end_to_end,
-    get_argument_parser,
-    print_rank_n
-)
+from utils import Model, print_rank_n
 
 
-class HFAccelerateModel(Model):
+class DSZeROModel(Model):
     def __init__(self, args: Namespace) -> None:
-        print_rank_n("Loading model...")
+        if (args.local_rank == 0):
+            print("Loading model...")
+
+        config = AutoConfig.from_pretrained(args.model_name)
+
+        world_size = int(os.getenv('WORLD_SIZE', '1'))
+        train_batch_size = 1 * world_size
+
+        ds_config = {
+            "fp16": {
+                "enabled": args.dtype == torch.float16,
+            },
+            "bf16": {
+                "enabled": args.dtype == torch.bfloat16,
+            },
+            "zero_optimization": {
+                "stage": 3,
+                "overlap_comm": True,
+                "contiguous_gradients": True,
+                "reduce_bucket_size": config.hidden_size * config.hidden_size,
+                "stage3_prefetch_bucket_size": 0.9 * config.hidden_size * config.hidden_size,
+                "stage3_param_persistence_threshold": 0
+            },
+            "steps_per_print": 2000,
+            "train_batch_size": train_batch_size,
+            "train_micro_batch_size_per_gpu": 1,
+            "wall_clock_breakdown": False
+        }
+
+        if (args.cpu_offload):
+            ds_config["zero_optimization"]["offload_param"] = {
+                "device": "cpu",
+                "pin_memory": True
+            }
+
+        # this tells from_pretrained to instantiate directly on gpus
+        dschf = HfDeepSpeedConfig(ds_config)
 
         self.tokenizer = AutoTokenizer.from_pretrained(args.model_name)
 
         self.model = AutoModelForCausalLM.from_pretrained(
-            args.model_name,
-            device_map="auto",
-            max_memory=get_max_memory_per_gpu_dict(
-                args.dtype, args.model_name),
-            torch_dtype=args.dtype
-        )
+            args.model_name, torch_dtype=args.dtype)
+        self.model = self.model.eval()
+        self.model = deepspeed.initialize(
+            model=self.model, config_params=ds_config)[0]
+        self.model.module.eval()
+        self.model = self.model.module
 
-        self.model.eval()
-        self.input_device = "cuda:0"
-
-        print_rank_n("Model loaded")
+        self.input_device = torch.cuda.current_device()
 
     def generate(self,
                  text: Union[str, List[str]],
@@ -64,18 +94,6 @@ class HFAccelerateModel(Model):
             output_tokens, skip_special_tokens=True)
 
         return output_text, generated_tokens
-
-
-def get_args():
-    parser = get_argument_parser()
-
-    group = parser.add_argument_group(title="launch config")
-    group.add_argument("--benchmark_cycles", type=int,
-                       default=0, help="additionally run benchmark")
-
-    args = utils.get_args(parser)
-
-    return args
 
 
 def get_max_memory_per_gpu_dict(dtype, model_name):
@@ -121,7 +139,3 @@ def get_max_memory_per_gpu_dict(dtype, model_name):
             f"Unable to generate the memory map automatically as the needed estimated memory per gpu ({param_memory_per_gpu_in_bytes/2**30:0.2f}GB) is bigger than the available per gpu memory ({max_memory_per_gpu_in_bytes/2**30:0.2f}GB)")
 
     return {i: param_memory_per_gpu_in_bytes for i in range(torch.cuda.device_count())}
-
-
-if (__name__ == "__main__"):
-    benchmark_end_to_end(get_args(), HFAccelerateModel)
