+++ inference/values.py
@@ -1,0 +1,3 @@
+HF_ACCELERATE = "hf_accelerate"
+DS_INFERENCE = "ds_inference"
+DS_ZERO = "ds_zero"
